backend/python/requirements.txt
# Build dependencies
setuptools>=65.0

# AI/ML Libraries
openai-whisper>=20231117
torch>=2.1.0
torchaudio>=2.1.0

# Audio Processing
pydub==0.25.1
librosa==0.10.1
soundfile==0.12.1

# Video Rendering
ffmpeg-python==0.2.0

# Web Framework (for API endpoints)
fastapi==0.109.0
uvicorn[standard]==0.26.0
pydantic==2.5.3

# Utilities
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
python-multipart==0.0.6

# Text Processing
rapidfuzz==3.6.1
